Add drush commands.
Add 2 drush commands:
ubc_prepare for preparing a CSV file based on another CSV file.
The output CSV file will contain only PIDS for the children.
ubc_generate for making the compounds based on a CSV file with only
PIDs for the children, and a PID (existing parent) or a label (new
parent) for the compound parent.
Refactored the code.